move | admin migrate command to | teragrep exec migrate
