+++ src/main/antlr4/com/teragrep/pth_03/antlr/DPLLexer.g4
@@ -46,7 +46,7 @@
 // -*- mode: conf; -*-
 lexer grammar DPLLexer;
 
-import GET_BOOLEAN, GET_INTEGER, GET_FIELD, GET_NUMBER, GET_INDEX_STRING, GET_REGEX_STRING, GET_SORT_FIELD, GET_TIME, GET_SPAN, GET_STRING, GET_WHERE_MODE, GET_WHERE_MODE_BY ,GET_T_OR_B, COMMAND_ADMIN_MODE, COMMAND_ABSTRACT_MODE, COMMAND_ACCUM_MODE, COMMAND_ADDCOLTOTALS_MODE, COMMAND_ADDINFO_MODE, COMMAND_ADDTOTALS_MODE, COMMAND_AGGREGATION_MODE, COMMAND_ANALYZEFIELDS_MODE, COMMAND_ANOMALIES_MODE, COMMAND_ANOMALOUSVALUE_MODE, COMMAND_ANOMALYDETECTION_MODE, COMMAND_APPENDCOLS_MODE, COMMAND_APPEND_MODE, COMMAND_APPENDPIPE_MODE, COMMAND_ARULES_MODE, COMMAND_ASSOCIATE_MODE, COMMAND_AUDIT_MODE, COMMAND_AUTOREGRESS_MODE, COMMAND_BIN_MODE, COMMAND_BUCKETDIR_MODE, COMMAND_CHART_MODE, COMMAND_CLUSTER_MODE, COMMAND_COFILTER_MODE, COMMAND_COLLAPSE_MODE, COMMAND_COLLECT_MODE, COMMAND_CONCURRENCY_MODE, COMMAND_CONTINGENCY_MODE, COMMAND_CONVERT_MODE, COMMAND_CORRELATE_MODE, COMMAND_DATAMODEL_MODE, COMMAND_DBINSPECT_MODE, COMMAND_DEDUP_MODE, COMMAND_DELETE_MODE, COMMAND_DELTA_MODE, COMMAND_DIFF_MODE, COMMAND_DPL_MODE, COMMAND_DUMP_MODE, COMMAND_EREX_MODE, COMMAND_EVAL_SUBMODE_BASE, EVAL_LANGUAGE_MODE, COMMAND_EVENTCOUNT_MODE, COMMAND_EVENTSTATS_MODE, COMMAND_EXPLAIN_MODE, COMMAND_EXTRACT_MODE, COMMAND_FIELDS_MODE, COMMAND_FIELDFORMAT_MODE, COMMAND_FIELDSUMMARY_MODE, COMMAND_FILLDOWN_MODE, COMMAND_FILLNULL_MODE, COMMAND_FINDKEYWORDS_MODE, COMMAND_FINDTYPES_MODE, COMMAND_FOLDERIZE_MODE, COMMAND_FOREACH_MODE, COMMAND_FORMAT_MODE, COMMAND_FROM_MODE, COMMAND_GAUGE_MODE, COMMAND_GENTIMES_MODE, COMMAND_GEOMFILTER_MODE, COMMAND_GEOM_MODE, COMMAND_GEOSTATS_MODE, COMMAND_HEAD_MODE, COMMAND_HIGHLIGHT_MODE, COMMAND_HISTORY_MODE, COMMAND_ICONIFY_MODE, COMMAND_INPUTCSV_MODE, COMMAND_INPUTLOOKUP_MODE, COMMAND_IPLOCATION_MODE, COMMAND_JOIN_MODE, COMMAND_KMEANS_MODE, COMMAND_KVFORM_MODE, COMMAND_LOADJOB_MODE, COMMAND_LOCALIZE_MODE, COMMAND_LOCALOP_MODE, COMMAND_LOOKUP_MODE, COMMAND_MAKECONTINUOUS_MODE, COMMAND_MAKEJSON_MODE, COMMAND_MAKEMV_MODE, COMMAND_MAKERESULTS_MODE, COMMAND_MAP_MODE, COMMAND_MCATALOG_MODE, COMMAND_MCOLLECT_MODE, COMMAND_METADATA_MODE, COMMAND_METASEARCH_MODE, COMMAND_MEVENTCOLLECT_MODE, COMMAND_MSTATS_MODE, COMMAND_MULTIKV_MODE, COMMAND_MULTISEARCH_MODE, COMMAND_MVCOMBINE_MODE, COMMAND_MVEXPAND_MODE, COMMAND_NOMV_MODE, COMMAND_OUTLIER_MODE, COMMAND_OUTPUTCSV_MODE, COMMAND_OUTPUTLOOKUP_MODE, COMMAND_OUTPUTTEXT_MODE, COMMAND_OVERLAP_MODE, COMMAND_PIVOT_MODE, COMMAND_PREDICT_MODE, COMMAND_RANGEMAP_MODE, COMMAND_RARE_MODE, COMMAND_REDISTRIBUTE_MODE, COMMAND_REGEX_MODE, COMMAND_RELEVANCY_MODE, COMMAND_RELTIME_MODE, COMMAND_RENAME_MODE, COMMAND_REPLACE_MODE, COMMAND_REST_MODE, COMMAND_RETURN_MODE, COMMAND_REVERSE_MODE, COMMAND_REX_MODE, COMMAND_REX4J_MODE, COMMAND_RTORDER_MODE, COMMAND_RUNSHELLSCRIPT_MODE, COMMAND_SAVEDSEARCH_MODE, COMMAND_SCRIPT_MODE, COMMAND_SCRUB_MODE, COMMAND_SEARCH_MODE, COMMAND_SEARCHTXN_MODE, COMMAND_SELFJOIN_MODE, COMMAND_SENDALERT_MODE, COMMAND_SENDEMAIL_MODE, COMMAND_SETFIELDS_MODE, COMMAND_SET_MODE, COMMAND_SICHART_MODE, COMMAND_SIRARE_MODE, COMMAND_SISTATS_MODE, COMMAND_SITIMECHART_MODE, COMMAND_SITOP_MODE, COMMAND_SORT_MODE, COMMAND_SPATH_MODE, COMMAND_STATS_MODE, COMMAND_STRCAT_MODE, COMMAND_STREAMSTATS_MODE, COMMAND_TABLE_MODE, COMMAND_TAGS_MODE, COMMAND_TAIL_MODE, COMMAND_TERAGREP_MODE, COMMAND_TIMECHART_MODE, COMMAND_TIMEWRAP_MODE, COMMAND_TOP_MODE, COMMAND_TRANSACTION_MODE, COMMAND_TRANSPOSE_MODE, COMMAND_TRENDLINE_MODE, COMMAND_TSCOLLECT_MODE, COMMAND_TSTATS_MODE, COMMAND_TYPEAHEAD_MODE, COMMAND_TYPELEARNER_MODE, COMMAND_TYPER_MODE, COMMAND_UNION_MODE, COMMAND_UNIQ_MODE, COMMAND_UNTABLE_MODE, COMMAND_WHERE_MODE, COMMAND_X11_MODE, COMMAND_XMLKV_MODE, COMMAND_XMLUNESCAPE_MODE, COMMAND_XPATH_MODE, COMMAND_XYSERIES_MODE;
+import GET_BOOLEAN, GET_INTEGER, GET_FIELD, GET_NUMBER, GET_INDEX_STRING, GET_REGEX_STRING, GET_SORT_FIELD, GET_TIME, GET_SPAN, GET_STRING, GET_WHERE_MODE, GET_WHERE_MODE_BY ,GET_T_OR_B, COMMAND_ABSTRACT_MODE, COMMAND_ACCUM_MODE, COMMAND_ADDCOLTOTALS_MODE, COMMAND_ADDINFO_MODE, COMMAND_ADDTOTALS_MODE, COMMAND_AGGREGATION_MODE, COMMAND_ANALYZEFIELDS_MODE, COMMAND_ANOMALIES_MODE, COMMAND_ANOMALOUSVALUE_MODE, COMMAND_ANOMALYDETECTION_MODE, COMMAND_APPENDCOLS_MODE, COMMAND_APPEND_MODE, COMMAND_APPENDPIPE_MODE, COMMAND_ARULES_MODE, COMMAND_ASSOCIATE_MODE, COMMAND_AUDIT_MODE, COMMAND_AUTOREGRESS_MODE, COMMAND_BIN_MODE, COMMAND_BUCKETDIR_MODE, COMMAND_CHART_MODE, COMMAND_CLUSTER_MODE, COMMAND_COFILTER_MODE, COMMAND_COLLAPSE_MODE, COMMAND_COLLECT_MODE, COMMAND_CONCURRENCY_MODE, COMMAND_CONTINGENCY_MODE, COMMAND_CONVERT_MODE, COMMAND_CORRELATE_MODE, COMMAND_DATAMODEL_MODE, COMMAND_DBINSPECT_MODE, COMMAND_DEDUP_MODE, COMMAND_DELETE_MODE, COMMAND_DELTA_MODE, COMMAND_DIFF_MODE, COMMAND_DPL_MODE, COMMAND_DUMP_MODE, COMMAND_EREX_MODE, COMMAND_EVAL_SUBMODE_BASE, EVAL_LANGUAGE_MODE, COMMAND_EVENTCOUNT_MODE, COMMAND_EVENTSTATS_MODE, COMMAND_EXPLAIN_MODE, COMMAND_EXTRACT_MODE, COMMAND_FIELDS_MODE, COMMAND_FIELDFORMAT_MODE, COMMAND_FIELDSUMMARY_MODE, COMMAND_FILLDOWN_MODE, COMMAND_FILLNULL_MODE, COMMAND_FINDKEYWORDS_MODE, COMMAND_FINDTYPES_MODE, COMMAND_FOLDERIZE_MODE, COMMAND_FOREACH_MODE, COMMAND_FORMAT_MODE, COMMAND_FROM_MODE, COMMAND_GAUGE_MODE, COMMAND_GENTIMES_MODE, COMMAND_GEOMFILTER_MODE, COMMAND_GEOM_MODE, COMMAND_GEOSTATS_MODE, COMMAND_HEAD_MODE, COMMAND_HIGHLIGHT_MODE, COMMAND_HISTORY_MODE, COMMAND_ICONIFY_MODE, COMMAND_INPUTCSV_MODE, COMMAND_INPUTLOOKUP_MODE, COMMAND_IPLOCATION_MODE, COMMAND_JOIN_MODE, COMMAND_KMEANS_MODE, COMMAND_KVFORM_MODE, COMMAND_LOADJOB_MODE, COMMAND_LOCALIZE_MODE, COMMAND_LOCALOP_MODE, COMMAND_LOOKUP_MODE, COMMAND_MAKECONTINUOUS_MODE, COMMAND_MAKEJSON_MODE, COMMAND_MAKEMV_MODE, COMMAND_MAKERESULTS_MODE, COMMAND_MAP_MODE, COMMAND_MCATALOG_MODE, COMMAND_MCOLLECT_MODE, COMMAND_METADATA_MODE, COMMAND_METASEARCH_MODE, COMMAND_MEVENTCOLLECT_MODE, COMMAND_MSTATS_MODE, COMMAND_MULTIKV_MODE, COMMAND_MULTISEARCH_MODE, COMMAND_MVCOMBINE_MODE, COMMAND_MVEXPAND_MODE, COMMAND_NOMV_MODE, COMMAND_OUTLIER_MODE, COMMAND_OUTPUTCSV_MODE, COMMAND_OUTPUTLOOKUP_MODE, COMMAND_OUTPUTTEXT_MODE, COMMAND_OVERLAP_MODE, COMMAND_PIVOT_MODE, COMMAND_PREDICT_MODE, COMMAND_RANGEMAP_MODE, COMMAND_RARE_MODE, COMMAND_REDISTRIBUTE_MODE, COMMAND_REGEX_MODE, COMMAND_RELEVANCY_MODE, COMMAND_RELTIME_MODE, COMMAND_RENAME_MODE, COMMAND_REPLACE_MODE, COMMAND_REST_MODE, COMMAND_RETURN_MODE, COMMAND_REVERSE_MODE, COMMAND_REX_MODE, COMMAND_REX4J_MODE, COMMAND_RTORDER_MODE, COMMAND_RUNSHELLSCRIPT_MODE, COMMAND_SAVEDSEARCH_MODE, COMMAND_SCRIPT_MODE, COMMAND_SCRUB_MODE, COMMAND_SEARCH_MODE, COMMAND_SEARCHTXN_MODE, COMMAND_SELFJOIN_MODE, COMMAND_SENDALERT_MODE, COMMAND_SENDEMAIL_MODE, COMMAND_SETFIELDS_MODE, COMMAND_SET_MODE, COMMAND_SICHART_MODE, COMMAND_SIRARE_MODE, COMMAND_SISTATS_MODE, COMMAND_SITIMECHART_MODE, COMMAND_SITOP_MODE, COMMAND_SORT_MODE, COMMAND_SPATH_MODE, COMMAND_STATS_MODE, COMMAND_STRCAT_MODE, COMMAND_STREAMSTATS_MODE, COMMAND_TABLE_MODE, COMMAND_TAGS_MODE, COMMAND_TAIL_MODE, COMMAND_TERAGREP_MODE, COMMAND_TIMECHART_MODE, COMMAND_TIMEWRAP_MODE, COMMAND_TOP_MODE, COMMAND_TRANSACTION_MODE, COMMAND_TRANSPOSE_MODE, COMMAND_TRENDLINE_MODE, COMMAND_TSCOLLECT_MODE, COMMAND_TSTATS_MODE, COMMAND_TYPEAHEAD_MODE, COMMAND_TYPELEARNER_MODE, COMMAND_TYPER_MODE, COMMAND_UNION_MODE, COMMAND_UNIQ_MODE, COMMAND_UNTABLE_MODE, COMMAND_WHERE_MODE, COMMAND_X11_MODE, COMMAND_XMLKV_MODE, COMMAND_XMLUNESCAPE_MODE, COMMAND_XPATH_MODE, COMMAND_XYSERIES_MODE;
 
 channels { DPLCOMMENT, ERRORCHANNEL }
 
@@ -270,7 +270,6 @@ mode COMMAND_MODE;
 COMMAND_MODE_SPACE: SPACE -> channel(HIDDEN);
 COMMAND_MODE_PIPE: '|' -> type(PIPE);
 
-COMMAND_MODE_ADMIN: 'admin' -> pushMode(COMMAND_ADMIN_MODE);
 COMMAND_MODE_ABSTRACT: 'abstract' -> pushMode(COMMAND_ABSTRACT_MODE);
 COMMAND_MODE_ACCUM: 'accum' -> pushMode(COMMAND_ACCUM_MODE), pushMode(GET_FIELD);
 COMMAND_MODE_ADDCOLTOTALS: 'addcoltotals' -> pushMode(COMMAND_ADDCOLTOTALS_MODE);
