+++ src/main/antlr4/imports/COMMAND_TERAGREP_MODE.g4
@@ -88,10 +88,12 @@ COMMAND_TERAGREP_MODE_LOAD: 'load';
 COMMAND_TERAGREP_MODE_SYSTEM: 'system';
 COMMAND_TERAGREP_MODE_VERSION: 'version';
 COMMAND_TERAGREP_MODE_ARCHIVE: 'archive';
+COMMAND_TERAGREP_MODE_MIGRATE: 'migrate' | 'MIGRATE';
+COMMAND_TERAGREP_MODE_EPOCH: 'epoch' | 'EPOCH';
 COMMAND_TERAGREP_MODE_SUMMARY: 'summary' -> pushMode(DEFAULT_MODE);
 COMMAND_TERAGREP_MODE_RETENTION: 'retention=' -> pushMode(GET_SPAN);
 COMMAND_TERAGREP_MODE_CODEC: 'codec=' -> pushMode(GET_STRING);
-COMMAND_TERAGREP_MODE_TABLE: 'table' -> pushMode(GET_FIELD);
+COMMAND_TERAGREP_MODE_TABLE: ('table' | 'TABLE') '='? -> pushMode(GET_FIELD);
 
 COMMAND_TERAGREP_MODE_HDFS_FORMAT: 'format=';
 COMMAND_TERAGREP_MODE_CSV_FORMAT: ('csv'|'CSV');
