+++ src/main/antlr4/imports/DPLParserTransform.g4
@@ -50,7 +50,6 @@ parser grammar DPLParserProcess;
 import DPLParserAggregations,
        DPLParserEvals,
        DPLParserTransform_abstract,
-       DPLParserTransform_admin,
        DPLParserTransform_accum,
        DPLParserTransform_addcoltotals,
        DPLParserTransform_addinfo,
@@ -208,7 +207,6 @@ import DPLParserAggregations,
 
 transformStatement
 	: abstractTransformation (PIPE transformStatement)?
-	    | adminTransformation (PIPE transformStatement)?
         | accumTransformation (PIPE transformStatement)?
         | addcoltotalsTransformation (PIPE transformStatement)?
         | addinfoTransformation (PIPE transformStatement)?
